refactor(train): S2 —— AnimaFamily 加载/前向收口 + registry 派发（多模型 PR-2b 阶段 2）
- families/protocol.py：ModelFamily 九方法契约（03 §4.1 冻结面 + 04 C4 的
  convert_lora_state_dict 恒等钩子）
- families/anima/loader.py：load_anima_model / load_text_encoders 自
  training/models.py 函数级迁入（族知识）；models.py 留 VAEWrapper +
  load_vae（跨族共享，D6）
- families/anima/forward.py：forward_with_optional_checkpoint 自
  model_loading.py 迁入（检查点展开是模型私货）；model_loading 留兼容
  re-export
- families/anima/family.py：AnimaFamily（load_dit 收编 attention_backend
  决策含 xformers 注入；encode_text_for_batch / forward_train 待 S3 接线）
- registry：get_family / resolve_family（惰性构造防环）；anima_train 公共
  命名空间 re-export 二者（D8' 派发咽喉）+ loader 来源改 families
- bootstrap 在 args 定稿后（pause snapshot / 交互补缺之后）resolve family
  fail-fast；phases/models.py 全部 load_* 经 ctx.family 派发；
  TrainingContext 增 family 字段

daemon parity 两条为已知顺序敏感存量问题（单独跑全绿）。
