+++ runtime/anima_train.py
@@ -86,9 +86,12 @@ from training.text_encoding import (  # noqa: E402
 from training.state import load_training_state, save_training_state  # noqa: E402
 from training.models import (  # noqa: E402
     ensure_models_namespace,
+    load_vae,
+)
+from training.families import get_family, resolve_family  # noqa: E402  # 派发咽喉（D8'）
+from training.families.anima.loader import (  # noqa: E402
     load_anima_model,
     load_text_encoders,
-    load_vae,
 )
 from training.sampling import sample_image  # noqa: E402
 from training.dataset import (  # noqa: E402
