+++ runtime/training/families/anima/forward.py
@@ -1,0 +1,37 @@
+"""Anima 族训练前向（多模型 PR-2b，自 training/model_loading.py 迁入）。
+
+forward_with_optional_checkpoint 手工展开 Anima 内部 API（prepare_embedded_sequence
+/ t_embedder / blocks / final_layer / unpatchify）做逐 block 梯度检查点 ——
+纯族知识（03 §4.3：检查点展开策略是模型私货）。
+"""
+
+from __future__ import annotations
+
+
+def forward_with_optional_checkpoint(model, latents, timesteps, cross, padding_mask, use_checkpoint=False):
+    """带可选梯度检查点的前向传播。"""
+    if not use_checkpoint:
+        return model(latents, timesteps, cross, padding_mask=padding_mask)
+    from torch.utils.checkpoint import checkpoint
+
+    x_B_T_H_W_D, rope_emb, extra_pos_emb = model.prepare_embedded_sequence(
+        latents, fps=None, padding_mask=padding_mask,
+    )
+    if timesteps.ndim == 1:
+        timesteps = timesteps.unsqueeze(1)
+    t_embedding, adaln_lora = model.t_embedder(timesteps)
+    t_embedding = model.t_embedding_norm(t_embedding)
+
+    block_kwargs = {
+        "rope_emb_L_1_1_D": rope_emb,
+        "adaln_lora_B_T_3D": adaln_lora,
+        "extra_per_block_pos_emb": extra_pos_emb,
+    }
+
+    for block in model.blocks:
+        def custom_forward(x, blk=block):
+            return blk(x, t_embedding, cross, **block_kwargs)
+        x_B_T_H_W_D = checkpoint(custom_forward, x_B_T_H_W_D, use_reentrant=False)
+
+    x_B_T_H_W_O = model.final_layer(x_B_T_H_W_D, t_embedding, adaln_lora_B_T_3D=adaln_lora)
+    return model.unpatchify(x_B_T_H_W_O)
