+++ runtime/training/model_loading.py
@@ -29,34 +29,6 @@ logger = logging.getLogger(__name__)
 # 梯度检查点
 # ============================================================================
 
-def forward_with_optional_checkpoint(model, latents, timesteps, cross, padding_mask, use_checkpoint=False):
-    """带可选梯度检查点的前向传播。"""
-    if not use_checkpoint:
-        return model(latents, timesteps, cross, padding_mask=padding_mask)
-    from torch.utils.checkpoint import checkpoint
-
-    x_B_T_H_W_D, rope_emb, extra_pos_emb = model.prepare_embedded_sequence(
-        latents, fps=None, padding_mask=padding_mask,
-    )
-    if timesteps.ndim == 1:
-        timesteps = timesteps.unsqueeze(1)
-    t_embedding, adaln_lora = model.t_embedder(timesteps)
-    t_embedding = model.t_embedding_norm(t_embedding)
-
-    block_kwargs = {
-        "rope_emb_L_1_1_D": rope_emb,
-        "adaln_lora_B_T_3D": adaln_lora,
-        "extra_per_block_pos_emb": extra_pos_emb,
-    }
-
-    for block in model.blocks:
-        def custom_forward(x, blk=block):
-            return blk(x, t_embedding, cross, **block_kwargs)
-        x_B_T_H_W_D = checkpoint(custom_forward, x_B_T_H_W_D, use_reentrant=False)
-
-    x_B_T_H_W_O = model.final_layer(x_B_T_H_W_D, t_embedding, adaln_lora_B_T_3D=adaln_lora)
-    return model.unpatchify(x_B_T_H_W_O)
-
 
 # ============================================================================
 # xformers 支持
@@ -283,3 +255,7 @@ def _load_weights_best_effort(model: torch.nn.Module, sd: dict, label: str) -> d
         "missing": missing,
         "unexpected": unexpected,
     }
+
+
+# 兼容 re-export（sister/loop 现有 import 面；多模型 PR-2b 移居 families/anima/forward.py）
+from training.families.anima.forward import forward_with_optional_checkpoint  # noqa: E402,F401
